requirements.txt
numpy
pygame
tensorflow-directml
numba